pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vercel"
version = "0.3.8"
description = "Python SDK for Vercel"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE", "LICENSE.*"]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "anyio>=4.0.0",
    "python-dotenv",
    "websockets>=12.0",
]

[project.scripts]
vercel = "vercel.__main__:main"
vercel-sandbox = "vercel.sandbox.__main__:main"
sandbox = "vercel.sandbox.__main__:main"

[dependency-groups]
dev = [
    "ruff>=0.5.0",
    "pytest>=7.0.0",
    "pytest-asyncio",
    "mypy",
    "build",
    "twine",
]

[tool.uv]
default-groups = ["dev"]
package = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["vercel*"]
namespaces = true

[tool.setuptools]
include-package-data = false

[tool.setuptools.package-data]
vercel = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
asyncio_mode = "auto"

[tool.mypy]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
isort = { combine-as-imports = true, known-first-party = ["vercel"] }

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = false

[tool.ruff.lint.per-file-ignores]
"examples/**/*.py" = ["E501"]
